requirements.txt
# Strands Agent Framework
strands-agents>=1.0.0
strands-agents-tools>=0.2.0

# AWS SDK
boto3>=1.35.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# JSON/Data Processing (built-in, no install needed)
# - json
# - datetime
# - typing
# - os
# - time
# - re
# - uuid
# - base64
# - sys
# - signal